+++ hppcrt/src/main/templates/com/carrotsearch/hppcrt/maps/KTypeVTypeHashMap.java
@@ -49,7 +49,7 @@ import com.carrotsearch.hppcrt.hash.*;
  */
 /*! ${TemplateOptions.generatedAnnotation} !*/
 public class KTypeVTypeHashMap<KType, VType>
-implements KTypeVTypeMap<KType, VType>, Cloneable
+        implements KTypeVTypeMap<KType, VType>, Cloneable
 {
     protected VType defaultValue = Intrinsics.<VType> empty();
 
@@ -64,8 +64,8 @@ implements KTypeVTypeMap<KType, VType>, Cloneable
           KType []
           #else !*/
     Object[]
-            /*! #end !*/
-            keys;
+    /*! #end !*/
+    keys;
 
     /**
      * Hash-indexed array holding all values associated to the keys.
@@ -75,8 +75,8 @@ implements KTypeVTypeMap<KType, VType>, Cloneable
           VType []
           #else !*/
     Object[]
-            /*! #end !*/
-            values;
+    /*! #end !*/
+    values;
 
     /*! #if ($RH) !*/
     /**
@@ -1741,7 +1741,7 @@ implements KTypeVTypeMap<KType, VType>, Cloneable
      * (constructor shortcut).
      */
     public static <KType, VType> KTypeVTypeHashMap<KType, VType> newInstance(final int initialCapacity,
-            final float loadFactor) {
+            final double loadFactor) {
         return new KTypeVTypeHashMap<KType, VType>(initialCapacity, loadFactor);
     }
