+++ hppcrt/src/main/templates/com/carrotsearch/hppcrt/sets/KTypeHashSet.java
@@ -46,8 +46,8 @@ import com.carrotsearch.hppcrt.hash.*;
  */
 /*! ${TemplateOptions.generatedAnnotation} !*/
 public class KTypeHashSet<KType>
-extends AbstractKTypeCollection<KType>
-implements KTypeLookupContainer<KType>, KTypeSet<KType>, Cloneable
+        extends AbstractKTypeCollection<KType>
+        implements KTypeLookupContainer<KType>, KTypeSet<KType>, Cloneable
 {
     /**
      * Hash-indexed array holding all set entries.
@@ -61,8 +61,8 @@ implements KTypeLookupContainer<KType>, KTypeSet<KType>, Cloneable
           KType []
           #else !*/
     Object[]
-            /*! #end !*/
-            keys;
+    /*! #end !*/
+    keys;
 
     /*! #if ($RH) !*/
     /**
@@ -966,7 +966,7 @@ implements KTypeLookupContainer<KType>, KTypeSet<KType>, Cloneable
      * Returns a new object of this class with no need to declare generic type (shortcut
      * instead of using a constructor).
      */
-    public static <KType> KTypeHashSet<KType> newInstance(final int initialCapacity, final float loadFactor) {
+    public static <KType> KTypeHashSet<KType> newInstance(final int initialCapacity, final double loadFactor) {
         return new KTypeHashSet<KType>(initialCapacity, loadFactor);
     }
